backend/requirements.txt
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
cachetools==5.5.2
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
google-auth==2.39.0
gunicorn==21.2.0
ics==0.7.2
idna==3.10
iniconfig==2.1.0
kombu==5.5.3
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
TatSu==5.13.1
types-python-dateutil==2.9.0.20241206
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
